Support adding groups as reviewer by SSH command
With the modify-reviewers SSH command it is now possible to also add
groups as reviewer for a change.

There are 2 Gerrit configuration parameter that limit the number of
reviewers that can be added at once by adding a group as reviewer.
1. addreviewer.maxAllowed
2. addreviewer.maxWithoutConfirmation
When adding a group as reviewer with the modify-reviewers command it
is ensured that the group that is added as reviewer has no more than
the configured addreviewer.maxAllowed members. However there is no
confirmation if a group has more members than the configured
addreviewer.maxWithoutConfirmation. The reason is that it is difficult
on the command line to ask for a confirmation and there is also less
danger than in the webui of accidentally choosing a wrong group.

Removing reviewers with the modify-reviewers command still does only
work for users and not for groups.

Change-Id: I76f6dbfd29ee2063b8b258710a29cba3db1ac035
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>
Bug: issue 881

Former-commit-id: 5e65d9b2d53543ff58bdb96f905043fbb9d61c0c