Apply states for projects (active, readonly and hidden)
This feature is the result of the discussion
about issue 349 which the project can assume
one of the states mentioned above.

Active state indicates the project is regular
and is the default value.

Read Only means that users can see the project
if read permission is granted, but all modification
operations are disabled.

Hidden means the project is not visible for
those who are not owners

Bug: issue 349
Change-Id: I2a8972701359c3792331712efc93cb6ded2065f5

Former-commit-id: acbf134d548fe17d7cfb797535e117226f6c6fb0