Predicates to check commit messages and edits
Adds predicate commit_message that returns the commit message as a symbol.
commit_message_matches takes in a regex pattern and checks it against
the commit message, returns true if a match is found.
Adds predicate commit_edits that takes in a regex pattern for filenames
and a regex pattern for edits. For all files in a commit that match
the filename regex, if the edits in any of those files match the
edit regex, then the predicate returns true.

Change-Id: I1c0b5ddb669aaca77908e18d7bb314c5aa6aec70

Former-commit-id: 0cec9e6f7ff1e62ce7f4e487e6a06e2a973baeb7