Fix build in m2eclipse 1.0
In version 1.0 of the eclipse maven plugin m2eclipse the project build
lifecycle is checked more strictly in order to avoid concurrency issues
[1]. Therefore the build failed. This commit fixes it by ignoring the
antrun and the build-helper-maven-plugin tasks in m2eclipse. This allows
for working in eclipse with m2eclipse 1.0 and does not harm the maven
build on the comand line.

[1] http://wiki.eclipse.org/M2E_plugin_execution_not_covered

Change-Id: Ic8324920e14b3b41b27e63b1820cf9d86e512a6a
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Former-commit-id: 2cb600d3d52a3dcff8a4382b089447c0f19022a1