Delete draft changes and patchsets
Adds ability to delete draft changes and patchsets that are not meant
or fit for code review. Deleting a draft patchset also deletes the
corresponding ref from the repository and decrements the next patch
set number for the change if necessary. Deleting a draft change
deletes all of its (draft) patchsets.

Change-Id: I04abeb67b64dd2366514e74d23f83066d409904e

Former-commit-id: 5a59261dc3c029e2fb3e08d3a8c6fc9e4dbce481