Verify the case of the project name before opening git repository
When a project is not in the project cache, Gerrit checks for the
corresponding git repository in the file system and adds the project
to the cache if the git repository is found. On Windows the paths are
case-insensitive, which means that the lookup for the git repository
in the file system is even successfull if the name of the project was
given in an incorrect case.

E.g. lets assume the project 'test/myProject' exists, then asking the
project cache for 'test/myProject', 'TEST/myProject', 'test/MyProject'
would all succeed. Even worse for each variant of the case a new
instance of the project would be added to the cache. These instances
of the same project in the project cache might even become
inconsistent with each other.

This problem is relevant for all SSH commands where the user can
specify a project name, e.g. for 'set-project-parent'. If for the
'--parent' parameter the project is specified in an incorrect case,
this incorrect case is then even persisted in the 'project.config'
file of the child projects.

To avoid these problems, with this change the case of the project name
is now verified and the project is only added to the cache if the
project name was specified in the correct case. If the project name
was specified in an incorrect case the lookup would fail with an
RepositoryNotFoundException.

Change-Id: Icf826d16e05a8c19537e86fe302b158bcb407bba
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: a7e928daaf7ba4176b5044797f90b21f420ab2f7