Remove unneeded escape characters from the documentation
There was an incompatible upgrade of asciidoc. The old version of
asciidoc required certain characters to be escaped with a backslash
and when the upgrade to the new version was done all those backslashes
that were used for escaping became visible. These unneeded backslashes
in the documenation are confusing to users, especially since they
mainly appear in syntax examples (where the syntax gets incorrect if
the backslahes are included). With this change all unneded backslashes
are removed from the Gerrit documentation.

Change-Id: I6b71ac558344b07cdbc73c67c472c0186e8f50f2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: a2bf0836a0e3dfdfd6e138ac9e0d5dcfa3b6ef89