Support auth.type = CUSTOM_EXTENSION
This new auth type configures the web UI similar to the way HTTP
works, allowing the UI to redirect sign-in requests through the
/login/* URL and sign-out requests through /logout.

No URL handlers are installed in the application for this type
of authentication system. Instead it is assumed additional code
has been injected into the same environment via Guice to handle
the authentication. This is currently a very advanced usage of
Gerrit Code Review's server and is not recommended for most
site administrators, so the type is not documented at this time.

This change is a first step towards supporting other types
of user authentication, with the idea of eventually having
a plugin system to permit other third party authenticators.

Change-Id: I95fcbfc6f486513f7c7105a1b7005ab78b1f4073
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 7bec78aa47a0f2bfd5b1bf35bd41223b52e197b3