ls-projects: new option to list all accessible projects
Add a new option '--all' to the 'ls-projects' SSH command to display
all projects that are accessible by the calling user account. Besides
the projects that the calling user account has been granted 'READ'
access to, this includes all projects that are owned by the calling
user account (even if for these projects the 'READ' access right is
not assigned to the calling user account).

Change-Id: I7e758e49ebff878a37f1fc550cbb63a99ea81818
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: f179be219f8306d1c7b8c6f031d94279db503fac