Support to reparent all children of a parent project
This change adds a new option to the 'set-project-parent' command that
allows to reparent all child projects of one parent project to another
parent project. It is possible to specify certain children of the old parent
that should not be reparented.

This is e.g. useful if you have defined a set of default access rights on a
parent project P and now one of its child projects C should not inherit these
default access rights. The new functionality allows you to easily introduce a
new hierarchy level under P:
1. create new child project D under P which defines the default access rights
2. reparent all child projects of P to D, except C (and D)
3. remove the default access rights from P

Change-Id: I3c6d8407f357f601c8000e6bc5cbc696012ecd09
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 04388419981597c17a1fe7c8a6e4e1975d989b94