Fix change state in patch set approval if reviewer is added to closed change
If a reviewer is added to a change, a dummy patch set approval is created
for the reviewer so that the change appears in the reviewer's dashboard.
In this patch set approval the change state (whether the change is open) is
cached. For the dummy patch set approval that is created when a reviewer is
added the cached change state is always open, which is incorrect if a
reviewer is added to a closed change. As a result the closed change will
appear in the reviewers dashboard in the 'Review Requests' section and will
stay there forever.

This change ensures that the correct change state is cached in the dummy
patch set approval when it is created.

Bug: issue 1139
Change-Id: I74ee137045010fa0d674328e2179e5031d9c42ec
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 3f5e1c46c3245a29a7eb2dafd2609a7a67e4a781