Support gwtorm 1.2
In 1.2 gwtorm requires accessor @Query methods to be unique
names within the interface. This means the arguments must not be
overloaded, and required fixing up a few interfaces that relied on
method overloading to select the right query statement.

1.2 also requires a unique id number for each relation in the
schema interface, similar to how earlier versions have a unique
column number in each entity. These can be used by NoSQL based
systems for unique identity, specially the protobuf backend that
exists to support the NoSQL backends of gwtorm.

Change-Id: I19421b966e336ee0e89e6743d79e989860d8c470

Former-commit-id: ae122ef21d2d6c1434b93184de598fae5010dea4