Embed the XSRF token in the host page again
If the servlet container forces 'HttpOnly' on us for our account
session cookie, we can't use that to prevent cross site forgery.
Instead embed a token that is unique to this session into the host
page, and have the web UI echo that token back on each request.
We'll validate the token matches the session cookie on the server,
and then simply never rotate it within the lifespan of the session.

Change-Id: Ia9678335b7446eab8a6ee7f043e03f928707b1ad
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a6b36ae97fbcb4cf30b59e94c01783288970ff29