Fix calculation of project name if repo is not existing
The calculation of the project name in LocalDiskRepositoryManager
assumes that the relative path to the git repository ends with a
slash. This is only true if the repository exists in the file system.
If a project inherits from a non existing parent (the inheritFrom in
project.config points to a non existing parent), we try to lookup a
non existing repository. In this case the relative path is just the
project name without any '/' in the end. As result the code fails
with a StringIndexOutOfBoundsException.

Change-Id: Id48c0708d077a75adc6f9ab9dbd91d051959d43c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: b57de70331f960286cc427d3e5201bcfe6922be0