Option for SSHD review-cmd to always publish the message.
"--force-message" option for the SSHD review command,
which allows Gerrit to publish the "--message", even if the
labels could not be applied due to change being closed.

Used by some scripts/CI-systems, where the results (or links
to the result) are posted as a message after completion of a
build (often together with a label-change, indicating the
success of the build).

If the message is posted successfully, the cmd will return
successfully, even if the label could not be changed.

Change-Id: Ic2d5fc25ba97ddcedce832c7444e4f07512c761c

Former-commit-id: 1fbf2c28a0c9115a7953a64535f91c747bf9abe0