Evict project in user cache on save of project meta data
On saving of the project description and project options the project
is not evicted from the user cache (PerRequestProjectControlCache).
As a result the ProjectInfoScreen is showing the old data after save.
Only when the page is reloaded the new data is displayed.

This change evicts the project from the user cache when the project
meta data is changed, so that the ProjectInfoScreen always shows the
correct data.

Bug: issue 1161
Change-Id: I6b23d73c0befdb051d6bde9a8ad309028e04d3ae
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 5746c35592933c71b539d8189674c845faf77638