Display proper error message when LDAP is unavailable
At the moment the error message 'Incorrect username or password.' is
displayed when querying the LDAP server fails because it is
unavailable. This message is misleading since the user cannot make
the login succeed by correcting the username or password. With this
change now a proper error message is displayed so that the user can
understand that login is not possible due to the LDAP server not
being available.

Change-Id: I4afa503642537dbdb33c1031220281f923dddfc8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 093eb7205617642b74153324442cd59eaf1df374