Add user pref to enable manual reviewing
Add a checkbox to the preferences header on the diff
screen which allows a user to specify whether they
want manual-reviewing enabled or disabled.  Previously,
every file was auto marked reviewed when a user first
displayed it.  The new manual mode prevents this auto
marking and only marks a file reviewed when the user
explicitly clicks on the reviewed button.

Using the manual mode allows users to better decide
when they feel they have reviewed a file.  This mode
is particularly useful for larger reviews when a
user may need to jump around and look at many other
files while commenting on a specific one.  This
prevents their jumping around from file to file from
marking the files as reviewed.  The intention is to
allow the reviewed status to provide more meaning
than simply "I have seen the file" (which in many
cases might even be a stretch).

Change-Id: Iaedbcd717ef8a587a78c56e3d1b5f3fcc86a6d21

Former-commit-id: e5490f1f5ad8be241a7888ec2199e4ed589386d4