Avoid double key event for GroupReferenceBox
When a key is pressed in the GroupReferenceBox the KeyPressHandler
which is registered on the SuggestBox gets the key event twice.
This is due to a bug in gwt:
http://code.google.com/p/google-web-toolkit/issues/detail?id=3533

Instead of adding the KeyPressHandler directly to the SuggestBox, it
should be added to the surrounding TextBoxBase. Then there is only
one event for every key press.

Change-Id: I09389ca77106a4e63f83c7a9c2d7b4a5ec99a458
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: dc422f33806c7841d586224e39c42f816de25326