set-project-parent: if update of 1 project fails continue with others
The set-project-parent SSH command can be used to update the parent
project of many child projects at once. If the update for one of the
child projects fails the command is interrupted and the further child
projects are not updated anymore.

With this change the error is now logged and printed out, but it does
not prevent the update of the further child projects.

Change-Id: Ibacf57b27dd2030ca4fea034d55b1921de400898
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: e21a566274cccacb9dd8dae9ce7902e91ea1a770