Detection of last node in TreeFormatter should not rely on object identity
If the nodes implement the equals() method the detection of the last
node may be inaccurate (e.g. if the node set contains two nodes for
which equals() returns true and one of them is the last one in the
sorted set). In this case a wrong node prefix would be printed out.
With this change the last node is now detected by counting the nodes
so that there is no dependency on the object identity anymore.

Change-Id: I533610780ac58d79e879475e9dc9f7ef7c5a04c2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 22cdc625699ca07a5033227a0fd3321b481b3d79