TreeFormatter should always use '\n' as line separator string
Currently the TreeFormatter is using 'PrintWriter#println()' to
write the line separator string to terminate the current line. This
method prints the line separator string that is defined by the system
property 'line.separator'. This line separator string may be
different from '\n' (e.g. on Windows it's '\r\n').

Since the formatting of the tree printed out by the TreeFormatter
should be always the same this change ensures that the TreeFormatter
now always prints '\n' as line separator string.

Change-Id: I1d1959ecba0786cf2ed58feebba7cfcbf18ac062
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: f0d93fb8146fa915d32367597ecbad07f96b03b8