Create and use a StarCache in the ChangeScreen
A StarCache keeps track of the starred state of a
Change and can inform its listeners when it changes.
It is also able to create star Images and KeyCommands
which already know how to update the StarCache, the
Images will also update themselves to reflect the
proper starred state of the Change they are bound to.

Use the StarCache in the ChangeScreen instead of
creating the star Image and KeyCommand and getting
and updating the starred value, and updating the star
Image.

Change-Id: I3f47fb1153674c6a172bf3319ad2d39df3762138

Former-commit-id: e52462edaf680ccdfb92dda65d6b0bc76c8013eb