Add better explanations to rejection messages
Gerrit currently rejects pushes by ambiguous messages e.g.
'fast-forward rejected', 'can not delete references'.
This is really confusing for people who have never used
gerrit and even for the administrators that don't know
about the access rights.

This patch aims to resolve this issue by telling why we
reject a certain 'Push' and how the user can get it fixed.

Ideally the MESSAGE_FOOTER would be read from a properties
file, but this still requires a total rebuild since the
strings file we have resides in the war package.

Change-Id: I270f680fd4a3d7b87aa3609926223c1759f66646

Former-commit-id: ece1a2ac34ce357b3e459f18a6b87f88a6727506