Allow loading Project Access when there is no refs/meta/config
If the project does not have refs/meta/config the revision is
null. Enable loading the access screen anyway with a null revision
field, and on save of any edits require the branch to be new.
This is already handled by MetaDataUpdate, provided that the base
revision we supply is null, so make it @Nullable inside of the
ChangeProjectAccess handler.

Change-Id: I4417bf802e329eb7193a7cb514f7ebcd7d68f26e

Former-commit-id: 23f2e7096ccf42ce834c0036cbdebb4129bebb84