Fixing asciidoc makefile to include images.
With the next release of Gerrit there will be images in the documentation.
With the current makefile they are not included. This patch will solve
this problem and include them.

With the advent of Internet Explorer 8 all major web browsers now support
the data URI scheme for embedded images and this patch use this feature.

Bug: issue 1231
Change-Id: I4e675ae42105f38c6ef1287621ac7d738930c82e

Former-commit-id: 21330d5d2b811ae44515a9c02b894588d460f812