Fix git push authentication over HTTP
Dynamically look up the request's CurrentUser each time it is
requested inside of Guice. It is possible for the session user
to change when the ProjectDigestFilter authenticates, as the
current session changes from anonymous to being identified.

The fact that this is a bug implies someone is asking for the
CurrentUser object earlier in the filter pipeline than we used
to be seeing. If it is before ProjectDigestFilter the result is
the AnonymousUser, which is incorrect.

Change-Id: I702c56237a54e7061b780cd5e5b419ef494b6d51

Former-commit-id: 62251f74ea0a4b7c4df71e5dd1134912cabb21b1