Improve error handling on branch creation
Improve the error messages that are displayed in the WebUI if the
creation of a branch fails due to invalid user input. The displayed
error messages are now coming from a properties file, which means
that they are translatable.

If the branch creation fails with LOCK_FAILURE because the branch
already exists or because the new branch should be created under an
existing branch (e.g. the users tries to create 'refs/heads/foo/bar'
when 'refs/heads/foo' already exists) now a human-readable error
message is displayed.

Change-Id: Iebc54a532863daeed3b77dfd941b35f89cf95bec
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 144f8dfd14652327ff1ac9a803882636681cce34